+++ frontend/src/app/main/ui/workspace/viewport/outline.cljs
@@ -67,7 +67,7 @@
   (let [shapes (obj/get props "shapes")
         zoom   (obj/get props "zoom")
         color  (if (or (> (count shapes) 1) (nil? (:shape-ref (first shapes))))
-                 "#31EFB8" "#00E0FF")]
+                 "var(--color-primary)" "var(--color-component-highlight)")]
     (for [shape shapes]
       [:& outline {:key (str "outline-" (:id shape))
                    :shape (gsh/transform-shape shape)