+++ frontend/src/app/main/ui/workspace/viewport/snap_points.cljs
@@ -15,7 +15,7 @@
    [beicon.core :as rx]
    [rumext.alpha :as mf]))
 
-(def ^:private line-color "#D383DA")
+(def ^:private line-color "var(--color-snap)")
 (def ^:private line-opacity 0.6)
 (def ^:private line-width 1)